# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flight Advisor Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "jstojiljkovic941@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/flight-advisor/v1/comments/{commentId}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Deletes comment",
                "description": "",
                "operationId": "commentDelete",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "Comment id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Remove the specified resource from storage.",
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when comment with the provided id does not exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Remove the specified resource from storage.",
                                            "type": "string",
                                            "example": "Comment with the provided id does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "City"
                ],
                "summary": "Updates an existing comment",
                "description": "",
                "operationId": "commentUpdate",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "Comment id to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update comment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "string",
                                        "example": "Serbia is a drinking problem country"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Update the specified resource in storage.",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Update the specified resource in storage.",
                                                        "type": "int",
                                                        "example": "1"
                                                    },
                                                    "comment": {
                                                        "description": "Update the specified resource in storage.",
                                                        "type": "string",
                                                        "example": "Serbia is a drinking problem country"
                                                    },
                                                    "city_id": {
                                                        "description": "Update the specified resource in storage.",
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "description": "Initial creation timestamp",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "readOnly": "true"
                                                    },
                                                    "updated_at": {
                                                        "description": "Last update timestamp",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "readOnly": "true"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Returns when validation fails",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Update the specified resource in storage.",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Update the specified resource in storage.",
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when comment with the provided id does not exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Update the specified resource in storage.",
                                            "type": "string",
                                            "example": "Comment with the provided id does not exist."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}